requirements.txt
openai
pinecone
langchain
langchain-openai
langchain-community
pandas
matplotlib
seaborn
numpy
pytest
python-dotenv
fastapi
uvicorn